agentcore/pyproject.toml
[project]
name = "agentcore"
version = "0.1.0"
description = "Intelligent agent for automatic code refactoring and integration across GitHub repositories."
readme = "README.md"
classifiers = [
    # TODO
]
license = {text = "MIT"}
dependencies = [
  "typer",
  "fastapi[standard]",
  "loguru",
  "pydantic",
  "pydantic-settings",
  "ipykernel",
  "reflex",
  "gitpython",
  "httpx",
]
requires-python = ">= 3.10"

[dependency-groups]
dev = [
    "coverage",
    "pre-commit",
    "pytest",
    "ruff",
]

[project.scripts]
repofactor = "repofactor.cli:app"

[tool.ty]
# All rules are enabled as "error" by default; no need to specify unless overriding.
# Example override: relax a rule for the entire project (uncomment if needed).
# rules.TY015 = "warn"  # For invalid-argument-type, warn instead of error.

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

[tool.uv]
package = true
